requirements.txt
bs4==0.0.1
earthengine-api==0.1.317
geopandas==0.9.0
matplotlib==3.5.2
ohsome==0.1.0
pandas==1.5.2
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
shapely==1.8.2
tensorflow==2.9.1
torch==1.12.0
tqdm==4.61.2
world-bank-data==0.1.3
bidi==0.22.0